requirements.txt
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
DateTime==5.3
decorator==5.1.1
executing==2.0.1
idna==3.6
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
zope.interface==6.1
